phy: keystone: updates to 10gbe gbe pcie serdes configurations
The updates in this patch mainly concerns improving the serdes
lane stability in the SerDes used by 10gbe, gbe and PCIe. In
order to work with these updates, there is also a corresponding
binary firmware update needed.

After these updates, the configurations are now divided mainly
in two parts, namely, a common configuration which is done
regardless a SerDes lane is actually used or not, and a lane
specific configuration which is done when a lane is actually
used by, for example, a network interface.

The common configuration is performed by the SerDes provider
driver, ie., phy-keystone-serdes.c, at driver probe. The lane
specific configuration is invoked by the corresponding peripheral
driver, for example network switch driver, through the PHY API.

The updates have been approved for release by 3rd party under
BSD license.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>