phy: ti-pipe3: Add set_mode callback to configure usb3 phy as pcie phy
DRA72 platform has the second instance of PHY shared between USB3
controller and PCIe controller with default as USB3 controller.
Since it is used with USB3 controller by default, it uses the
compatible specific to USB (ti,omap-usb3).

Populate set_mode callback so that the USB3 PHY can be configured
to be used with PCIe controller.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>